Explicitly reserve memory regions based on the input E820 map
The input E820 map now marks the memory regions that need to be
reserved.

Mark them explicitly when publishing memory regions in PlatformPei,
except when they are in the [PcdPciExpressBaseAddress, 4GB) area.
We do not allow the input E820 map to reserve memory in this area
because PlatformPei and AcrnEmuVariableFvbRuntimeDxe make certain
assumptions about it.

Due to this design change, the 64-bit PCI host aperture is now
hard-coded.

Signed-off-by: Peter Fang <peter.fang@intel.com>